test(sparkly-client, sparkly-example): move updater client mock code to package
